Move audio and mwscript-related patches to their own files
