+++ MWSE/PatchAudio.h
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace mwse::patch::audio {
+	void install();
+	void installPostLua();
+}
